// src/app/models/cliente.ts
export class Cliente {
    IdCliente: number;
    Tipo: string;
    Nome: string;
    Sexo: string;
    Pai: string;
    Mae: string;
    Profissao: string;
    Site: string;
    Conjuge: string;
    CpfCnpj: string;
    DataNascimento: string;
    CPF: string;
    RazaoSocial: string;
    NomeFantasia: string;
    CNPJ: string;
    Rg: string;
    InscricaoEstadual: string;
    Telefone1: string;
    Telefone2: string;
    Email: string;
    Observacao: string;
    Situacao: string;
}